Refactor DB models, general code cleanup
