+++ app/controllers/user.py
@@ -1,7 +1,7 @@
 from flask import g
 
 from app.helpers.render import render_json, render_error
-from app.models import User
+from app.models.User import User
 
 
 def get_my_profile():
