+++ app/helpers/__init__.py
@@ -1,4 +1,1 @@
-__all__ = [
-    'render',
-    'tasks'
-]
+
