+++ app/models/Post.py
@@ -3,7 +3,6 @@ import datetime
 from sqlalchemy.dialects.mysql import LONGTEXT
 
 from app.instances.db import db
-from app.models.User import User
 from config import posts
 
 
@@ -12,14 +11,15 @@ class Post(db.Model):
     Represnts a post (e.g. challenge)
     """
 
+    __tablename__ = 'posts'
+
     id = db.Column(db.Integer, primary_key=True, unique=True, autoincrement=True)
     title = db.Column(db.String(posts['max_title']), nullable=False)
     body = db.Column(LONGTEXT, nullable=False)
 
     date_created = db.Column(db.DateTime, default=datetime.datetime.utcnow)
 
-    user_id = db.Column(db.Integer, db.ForeignKey(User.id), nullable=False)
-    user = db.relationship(User, backref=db.backref('posts'))
+    user_id = db.Column(db.Integer, db.ForeignKey('users.id'), nullable=False)
 
     def to_json(self):
         data = {}
