--- app/models/__init__.py
@@ -1,2 +1,0 @@
-from app.models.Post import Post
-from app.models.User import User, UserJWTToken
