+++ requirements.txt
@@ -11,3 +11,4 @@ webassets>=0.12.1
 markdown>=2.6.9
 bleach>=2.1.1
 pylibmc>=1.5.2
+SQLAlchemy>=1.1.15
