+++ static/.esdoc.json
@@ -48,7 +48,7 @@
       "name": "esdoc-importpath-plugin",
       "option": {
         "replaces": [
-          { "from": "^(\\./)?js", "to": "~" },
+          { "from": "^(\\./)?js", "to": "~" }
         ]
       }
     }
