+++ tests/user_tests/test_user_model.py
@@ -1,4 +1,4 @@
-from app.models import User
+from app.models.User import User
 from tests.test_base import TestBase
